refactor get_lean_repos to use the github_api module
