+++ src/sorryscraper/crawler/github_api.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, Set
 import requests
 from requests.adapters import HTTPAdapter
 from urllib3.util.retry import Retry
@@ -364,3 +364,51 @@ def get_contributors(repo: str, session: requests.Session) -> List[str]:
         contributors.add(owner)
         
     return sorted(contributors) 
+
+def get_user_repos(user: str, session: requests.Session) -> Set[str]:
+    """Get all non-fork repositories for a user.
+    
+    Args:
+        user: GitHub username
+        session: Authenticated GitHub session
+        
+    Returns:
+        Set of repository full names (owner/name)
+    """
+    repos = set()
+    page = 1
+    
+    while True:
+        check_rate_limit(session)
+        response = session.get(
+            f"https://api.github.com/users/{user}/repos",
+            params={"page": page, "per_page": 100, "type": "owner"}
+        )
+        response.raise_for_status()
+        
+        results = response.json()
+        if not results:
+            break
+            
+        for repo in results:
+            if not repo["fork"] and not repo["archived"]:
+                repos.add(repo["full_name"])
+        
+        page += 1
+    
+    return repos
+
+def has_lakefile(repo: str, session: requests.Session) -> bool:
+    """Check if a repository has a lakefile.lean or lakefile.toml.
+    
+    Args:
+        repo: Repository name in format 'owner/name'
+        session: Authenticated GitHub session
+        
+    Returns:
+        True if repository has either lakefile
+    """
+    check_rate_limit(session)
+    lean_response = session.get(f"https://api.github.com/repos/{repo}/contents/lakefile.lean")
+    toml_response = session.get(f"https://api.github.com/repos/{repo}/contents/lakefile.toml")
+    return lean_response.status_code == 200 or toml_response.status_code == 200
